scripts/requirements.txt
gensim==4.3.1
rich==13.5.2
pgvector==0.2.1
psycopg2-binary==2.9.7
